# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_your-org_radius-go_internal_domain.CoAChangeRequest:
    properties:
      bandwidth_max_down:
        type: integer
      bandwidth_max_up:
        type: integer
      max_total_octets:
        type: integer
      mikrotik_group:
        type: string
      rate_limit:
        type: string
      username:
        type: string
    type: object
  github_com_your-org_radius-go_internal_domain.CreateNASRequest:
    properties:
      description:
        type: string
      ip_address:
        type: string
      name:
        type: string
      secret:
        type: string
    type: object
  github_com_your-org_radius-go_internal_domain.CreateUserRequest:
    properties:
      bandwidth_max_down:
        type: integer
      bandwidth_max_up:
        type: integer
      email:
        type: string
      framed_ip:
        type: string
      full_name:
        type: string
      idle_timeout:
        type: integer
      max_total_octets:
        type: integer
      mikrotik_group:
        type: string
      password:
        type: string
      rate_limit:
        type: string
      service_type:
        type: string
      session_timeout:
        type: integer
      simultaneous_use:
        type: integer
      username:
        type: string
    type: object
...
  github_com_your-org_radius-go_internal_domain.DisconnectRequest:
    properties:
      reason:
        type: string
      username:
        type: string
    type: object
...
  github_com_your-org_radius-go_internal_domain.NAS:
    properties:
      created_at:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      ip_address:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  github_com_your-org_radius-go_internal_domain.RadiusSession:
    properties:
      bandwidth_max_down:
        type: integer
      bandwidth_max_up:
        type: integer
      called_station:
        type: string
      calling_station:
        type: string
      created_at:
        type: string
      framed_ip:
        type: string
      id:
        type: string
      input_octets:
        type: integer
      last_update:
        type: string
      max_total_octets:
        type: integer
      mikrotik_group:
        type: string
      nas_id:
        type: string
      nas_identifier:
        type: string
      nas_ip:
        type: string
      output_octets:
        type: integer
      rate_limit:
        type: string
      service_type:
        $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.ServiceType'
      session_id:
        type: string
      session_status:
        $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.SessionState'
      session_time:
        type: integer
      start_time:
        type: string
      stop_time:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  github_com_your-org_radius-go_internal_domain.ServiceType:
    enum:
    - framed
    - login
    type: string
    x-enum-varnames:
    - ServiceTypeFramed
    - ServiceTypeLogin
  github_com_your-org_radius-go_internal_domain.SessionState:
    enum:
    - active
    - stopped
    - stale
    type: string
    x-enum-varnames:
    - SessionStateActive
    - SessionStateStop
    - SessionStateStale
  github_com_your-org_radius-go_internal_domain.Subscriber:
    properties:
      bandwidth_max_down:
        type: integer
      bandwidth_max_up:
        type: integer
      created_at:
        type: string
      email:
        type: string
      enabled:
        type: boolean
      expires_at:
        type: string
      framed_ip:
        type: string
      full_name:
        type: string
      id:
        type: string
      idle_timeout:
        type: integer
      is_voucher:
        type: boolean
      max_total_octets:
        type: integer
      mikrotik_group:
        type: string
      rate_limit:
        type: string
      service_type:
        $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.ServiceType'
      session_timeout:
        type: integer
      simultaneous_use:
        type: integer
      updated_at:
        type: string
      username:
        type: string
      voucher_package_id:
        type: string
    type: object
...
  github_com_your-org_radius-go_internal_domain.UpdateNASRequest:
    properties:
      description:
        type: string
      enabled:
        type: boolean
      ip_address:
        type: string
      name:
        type: string
      secret:
        type: string
    type: object
  github_com_your-org_radius-go_internal_domain.UpdateUserRequest:
    properties:
      bandwidth_max_down:
        type: integer
      bandwidth_max_up:
        type: integer
      email:
        type: string
      enabled:
        type: boolean
      framed_ip:
        type: string
      full_name:
        type: string
      idle_timeout:
        type: integer
      max_total_octets:
        type: integer
      mikrotik_group:
        type: string
      password:
        type: string
      rate_limit:
        type: string
      service_type:
        type: string
      session_timeout:
        type: integer
      simultaneous_use:
        type: integer
      username:
        type: string
    type: object
...
  github_com_your-org_radius-go_internal_runtime.Envelope:
    properties:
      data: {}
      error: {}
      meta: {}
      success:
        type: boolean
    type: object
  internal_radius.CleanupResult:
    properties:
      active_sessions_kept:
        type: integer
      cleaned_at:
        type: string
      stale_sessions_cleaned:
        type: integer
    type: object
  internal_radius.CoaChangeResult:
    properties:
      disconnected_count:
        type: integer
      failed_nas:
        items:
          type: string
        type: array
    type: object
...
  internal_radius.Status:
    properties:
      acct_requests:
        type: integer
      active_sessions:
        type: integer
      auth_accepts:
        type: integer
      auth_rejects:
        type: integer
      auth_requests:
        type: integer
      health:
        type: string
      nas_count:
        type: integer
      started_at:
        type: string
      subscriber_count:
        type: integer
    type: object
host: localhost:8083
info:
  contact: {}
  description: |-
    RADIUS server with HTTP management API.
    Management endpoints require the X-Internal-Secret header.
  title: radius-go API
  version: "1.0"
paths:
  /radius/nases:
    get:
      description: List all configured Network Access Servers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.NAS'
                  type: array
              type: object
      security:
      - InternalSecret: []
      summary: List NAS devices
      tags:
      - NAS
    post:
      consumes:
      - application/json
      description: Register a new Network Access Server
      parameters:
      - description: NAS details
        in: body
        name: nas
        required: true
        schema:
          $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.CreateNASRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.NAS'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
      security:
      - InternalSecret: []
      summary: Create NAS device
      tags:
      - NAS
  /radius/nases/{id}:
    delete:
      parameters:
      - description: NAS UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
      security:
      - InternalSecret: []
      summary: Delete NAS device
      tags:
      - NAS
    put:
      consumes:
      - application/json
      parameters:
      - description: NAS UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated NAS fields
        in: body
        name: nas
        required: true
        schema:
          $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.UpdateNASRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.NAS'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
      security:
      - InternalSecret: []
      summary: Update NAS device
      tags:
      - NAS
  /radius/sessions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.RadiusSession'
                  type: array
              type: object
      security:
      - InternalSecret: []
      summary: List active sessions
      tags:
      - Sessions
  /radius/sessions/cleanup:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_radius.CleanupResult'
              type: object
      security:
      - InternalSecret: []
      summary: Cleanup stale sessions
      tags:
      - Sessions
  /radius/sessions/disconnect:
    post:
      consumes:
      - application/json
      parameters:
      - description: Disconnect request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.DisconnectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
      security:
      - InternalSecret: []
      summary: Disconnect user sessions (PoD)
      tags:
      - Sessions
  /radius/sessions/reconcile:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
      security:
      - InternalSecret: []
      summary: Reconcile sessions from DB into memory
      tags:
      - Sessions
  /radius/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_radius.Status'
              type: object
      security:
      - InternalSecret: []
      summary: Server status and counters
      tags:
      - Status
  /radius/subscribers:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.Subscriber'
                  type: array
              type: object
      security:
      - InternalSecret: []
      summary: List subscribers
      tags:
      - Subscribers
    post:
      consumes:
      - application/json
      parameters:
      - description: Subscriber fields
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.Subscriber'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
      security:
      - InternalSecret: []
      summary: Create subscriber
      tags:
      - Subscribers
  /radius/subscribers/{id}:
    delete:
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
      security:
      - InternalSecret: []
      summary: Delete subscriber
      tags:
      - Subscribers
    put:
      consumes:
      - application/json
      parameters:
      - description: User UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated fields
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.Subscriber'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
      security:
      - InternalSecret: []
      summary: Update subscriber
      tags:
      - Subscribers
  /radius/subscribers/coa-change:
    post:
      consumes:
      - application/json
      parameters:
      - description: CoA change request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_your-org_radius-go_internal_domain.CoAChangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_radius.CoaChangeResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_your-org_radius-go_internal_runtime.Envelope'
      security:
      - InternalSecret: []
      summary: Change user profile via CoA
      tags:
      - Sessions
...
schemes:
- http
securityDefinitions:
  InternalSecret:
    in: header
    name: X-Internal-Secret
    type: apiKey
swagger: "2.0"